refactor: separate plan system from mapsystem
