+++ Client/Assets/Scripts/Managers/Contents/MapManager.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 public class MapManager
 {
     public MapSystem MapSystem { get; set; }
+    public PlanSystem PlanSystem { get; set; }
 
     public void Init()
     {
